server/requirements.txt
# ====== Core Framework ======
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.0.0

# ====== AI / LLM (Gemini Only) ======
google-generativeai>=0.3.0

# ====== Question API (NLP) ======
scikit-learn>=1.3.0
numpy>=1.24.0

# ====== STT (Whisper) ======
faster-whisper>=0.10.0

# ====== PyTorch (for vllm/xformers) ======
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# ====== vLLM & xformers ======
vllm==0.10.2
xformers==0.0.32.post1

# ====== Voice Analysis ======
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
praat-parselmouth>=0.4.3

# ====== RAG Pipeline ======
chromadb>=0.4.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
rank-bm25>=0.2.2
tiktoken>=0.5.0

# ====== PDF Processing ======
unstructured>=0.11.0
unstructured[pdf]
pdf2image>=1.16.0
Pillow>=10.0.0

# ====== Utilities ======
requests>=2.31.0
aiofiles>=23.2.0